CommentList: add version with links
